### 2024-09-25 #213 @cklunch
neonutilities submission
Submitting Author: Claire Lunch (@cklunch)
All current maintainers: (@cklunch, @bhass-neon, @znickerson8)
Package Name: neonutilities
One-Line Description of Package: neonutilities is a package for accessing and wrangling data generated and published by the National Ecological Observatory Network.
Repository Link:  https://github.com/NEONScience/NEON-utilities-python
Version submitted: v1.0.1
EiC: TBD
Version accepted: TBD

---

The neonutilities Python package provides utilities for discovering, downloading, and working with data files published by the National Ecological Observatory Network (NEON). NEON data files can be downloaded from the [NEON Data Portal](http://data.neonscience.org/) or [API](http://data.neonscience.org/data-api). The neonutilities package includes wrapper functions for the API and functions to reformat and stack NEON tabular data for analysis. This is a Python-native adaptation of the heavily used [neonUtilities R package](https://cran.r-project.org/web/packages/neonUtilities/index.html).

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The target audience is scientists doing research using NEON data. The package enables programmatic workflows for data downloading, and provides a standardized way to merge the product-site-month data files NEON publishes, making them analysis-ready.

There is an incomplete package on PyPi [here](https://pypi.org/project/py-neonUtilities/) that was started in 2020 by a student at a coding camp. It doesn't appear to have been finished, and is not maintained. Some NEON users have developed their own code to do some of the functionality covered by neonutilities, but as far as I know none of them have shared it broadly.
