### 2024-09-30 #213 @cmarmo
Hi @cklunch! Thank you for submitting your package for pyOpenSci review.
Below are the basic checks that your package needs to pass to begin our review.
If some of these are missing, we will ask you to work on them before the review process begins.

Please check our [Python packaging guide](https://www.pyopensci.org/python-package-guide) for more information on the elements below.

---
---

Here my comments to clarify why I haven't ticked all the check-boxes.
- Documentation: I've found a lot of relevant information in the https://www.neonscience.org/ page, however there is no standalone documentation for the package, even if a `docs` directory is present in the repo with an empty `index.md`. I strongly recommend to add some documentation pages in order to structure the access to the information. There is no need to rewrite what is already on https://www.neonscience.org/, just linking and structuring, following for example the structure suggested in the pre-review checks. Also, it is not clear to me what is part of the Neon Data API and what is part of the python implementation.
- The README does not contain the documentation about how to contribute: it is true that the CONTRIBUTING file specifies that external contributions will not be considered for now. Perhaps a link to the CONTRIBUTING would be enough.
- As opening issues is the only way to contribute for now, it would be nice to have issue templates for different kind of contributions (bug reports, new feature suggestions, documentation,...)

I think the fluidity of the review process will benefit from those improvements.

Thank you so much for your understanding!
