### 2024-09-30 #213 @lwasser
hey there @cklunch @bhass-neon @znickerson8 👋🏻 it's nice to see you here on GitHub! I have a question for you about your contributing doc which says:

> The neonutilities package is currently not accepting external contributions. If you have a suggestion for a fix or enhancement to the package, please create an [Issue](https://github.com/NEONScience/NEON-utilities-python/issues) in this repository, or contact us using the NEON [Contact Us](https://www.neonscience.org/about/contact-us) page.

We do consider opening issues to report bugs and to request useful features to be contributions. They are a different type than a PR of course. What is the intent of "we are not accepting contributions" is in this case? We are having some discussions around our pyOpenSci policies, maintainer responsiveness to users and open source processes in general related to this package. Any input from you would be super helpful! thank you!!
