### 2024-10-01 #213 @cklunch
@cmarmo @lwasser Thanks for the feedback!

For documentation, the intent was for the tutorials that are linked in the readme to provide the documentation users need to understand how to use the package. The [Download and Explore](https://www.neonscience.org/resources/learning-hub/tutorials/download-explore-neon-data) tutorial provides instructions in the most commonly used functions, and some context and common follow-up data wrangling, and the [neonUtilities](https://www.neonscience.org/resources/learning-hub/tutorials/neondatastackr) tutorial provides a function index. If this isn't what you're looking for, or if different content is needed, do you have an example of documentation that would be appropriate?

For contributing, of course bug reports and requests are always welcome! We've found with the R package that it isn't realistic for external folks to contribute code directly. The package is so specific to the NEON publication system, it's hard for people who aren't deeply familiar with that system to write generic code for it. It's almost always easier for us to incorporate requested changes ourselves than to work with external PRs. But if there's a better way to express that in the documentation, I'm very happy to.

I'll get issue templates set up in the repo, and wait to update documentation based on what we decide in this discussion. Thanks!
