### 2024-10-03 #213 @cmarmo
> For documentation, the intent was for the tutorials that are linked in the readme to provide the documentation users need to understand how to use the package. The [Download and Explore](https://www.neonscience.org/resources/learning-hub/tutorials/download-explore-neon-data) tutorial provides instructions in the most commonly used functions, and some context and common follow-up data wrangling, and the [neonUtilities](https://www.neonscience.org/resources/learning-hub/tutorials/neondatastackr) tutorial provides a function index. If this isn't what you're looking for, or if different content is needed, do you have an example of documentation that would be appropriate?

Indeed, that's why I said that a lot of information is already there. The way it is structured though is not easy to browse for someone new to the neon ecosystem.
It would be easier also for the review, to have a reference documentation page (like your index.md), rendered by github pages or similar, with a table of contents and the essential information: Getting started, tutorials, API, and then the links to the related pages already available in the neon website.
Does that sound reasonable?
