### 2024-12-04 #213 @cklunch
@lwasser @cmarmo Changes to neonutilities have been pushed, including creating issue templates, clarifying in the CONTRIBUTING file, and creating a [ReadTheDocs](https://neon-utilities-python.readthedocs.io/en/latest/index.html) page. Thank you for your guidance!

https://github.com/NEONScience/NEON-utilities-python
