### 2024-12-07 #213 @cmarmo
Thank you for your work @cklunch !
I have checked the related items in my initial comment.... just a nitpick ... at the end of the [get started](https://neon-utilities-python.readthedocs.io/en/latest/getting-started.html) page the link to the tutorials is not linked....

I'm going to start looking for an editor.
