### 2025-02-12 #213 @lwasser
hey @cklunch my apologies for the delay! i am putting out a call for editors today on social (bluesky, fosstodon and linked in). we will need to onboard someone new. Please bare with us. The state of the world is hitting us all I think so things are moving more slowly now. I hope you and everyone at NEON are doing well.
