### 2025-03-14 #213 @JuliMillan
## Editor response to review:

---

## Editor comments
:wave: Hi @ethanwhite and @benjamindonnachie! Thank you for volunteering to review for pyOpenSci! It's great to have you both here.

2. Here is the **[review template](https://www.pyopensci.org/software-peer-review/how-to/reviewer-guide.html#peer-review-template)** that you will need to fill out and submit here as a comment, once your review is complete.

Please get in touch with any questions or concerns! Your review is due: April 4 2025
---

Reviewers: @ethanwhite, @benjamindonnachie
Due date: April 4 2025
