### 2025-03-16 17f3215
Potential revision to optimise convert_byte_size
I've been optimising a lot of Python code recently and offer this as a potential alternative to convert_byte_size().  Removing the nested if.. elif... statements and making it easier to add additional si prefixes in the future if desired.
