### 2025-03-18 NEONScience/NEON-utilities-python#15 @benjamindonnachie
requirements.txt does not include version numbers
Is there an opportunity to include version numbers for required packages?  Or perhaps minimum version for major libraries?

For example, Pandas is quite fast-moving, and although you don't seem to use anything contentious, the behaviour of pd.read_csv changed between v1 and v2.
