### 2025-03-20 7a82c76
Create files_by_uri.py
@cklunch This is the first draft of the new files_by_uri function. I ran an initial test of downloading all GEO URLs to a local directory. In the next week, I will develop more testing for the function. Feel free to review the function whenever as I work on testing.
