+++ src/neonutilities/files_by_uri.py
@@ -1,0 +1,174 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+
+import requests
+import os
+import platform
+import logging
+import importlib.metadata
+import pandas as pd
+from tqdm import tqdm
+from urllib.parse import urlparse
+from .unzip_and_stack import unzip_zipfile
+from .helper_mods.metadata_helpers import convert_byte_size 
+
+# Set global user agent
+vers = importlib.metadata.version('neonutilities')
+plat = platform.python_version()
+osplat = platform.platform()
+usera = f"neonutilities/{vers} Python/{plat} {osplat}"
+
+def files_by_uri(filepath,
+                 savepath=None,
+                 check_size=True,
+                 unzip=True,
+                 save_zipped_files=False,
+                 progress=True,
+                 ):
+    """
+
+    Get files from NEON GCS Bucket using URLs in stacked data
+    
+    Parameters
+    --------
+    filepath: The location of the NEON data containing URIs. Can be either a local directory containing NEON tabular data or a list object containing tabular data.
+    savepath: The location to save the output files from the GCS bucket, optional. Defaults to creating a "GCS_zipFiles" folder in the filepath directory.
+    check_size: should the user be told the total file size before downloading, defaults to True? (true/false)
+    unzip: indicates if the downloaded files from GCS buckets should be unzipped into the same directory, defaults to True. Supports .zip and .tar.gz files currently. (true/false)
+    save_zipped_files: Should the unzipped monthly data folders be retained, defaults to False? Supports .zip and .tar.gz files currently. (true/false)
+    progress: Should a progress bar be displayed?
+    
+    Return
+    --------
+    A folder in the working directory (or in savepath, if specified), containing all files meeting query criteria.
+    
+    Example
+    --------
+    ZN NOTE: Insert example when function is coded
+    
+    >>> example
+    
+    Created on Fri Aug 9 2024
+    
+    @author: Zachary Nickerson
+    """ 
+    
+    # check that filepath points to either a directory or a Python list object
+    if not isinstance(filepath, list):
+        if not os.path.exists(filepath):
+            raise Exception("Input filepath is not a list object in the environment nor an existing file directory.")
+            
+    # if filepath is a directory, read in contents
+    if isinstance(filepath, list):
+        tabList = filepath
+        if not os.path.exists(savepath):
+            try:
+                os.makedirs(savepath)
+            except OSError:
+                print(f"Could not create savepath directory. Files will be saved to {os.getcwd()}/GCS_zipFiles")
+                savepath = f"{os.getcwd()}/GCS_zipFiles"
+    else:
+        files = [os.path.join(filepath, f) for f in os.listdir(filepath) if os.path.isfile(os.path.join(filepath, f))]
+        tabList = {}
+        for j, file in enumerate(files):
+            try:
+                tabList[file] = pd.read_csv(file)
+            except Exception as e:
+                print(f"File {file} could not be read.")
+                tabList[f"error{j}"] = None
+                continue
+        tabList = {k: v for k, v in tabList.items() if not k.startswith('error')}
+        
+    # Check for the variables file in the filepath
+    varList = [k for k in tabList.keys() if "variables" in k]
+    if len(varList) == 0:
+        raise Exception("Variables file was not found in specified filepath.")
+    if len(varList) > 1:
+        raise Exception("More than one variables file found in filepath.")
+    varFile = tabList[varList[0]]
+    
+    URLs = varFile[varFile['dataType'] == 'uri']
+    
+    # All of the tables in the package with URLs to download
+    allTables = URLs['table'].unique()
+    
+    # Loop through tables and fields to compile a list of URLs to download
+    URLsToDownload = []
+    
+    # Remove allTables values that aren't in tabList
[…]
