### 2025-03-26 #213 @ethanwhite
- [x] **Installation instructions:** for the development vers

Package not submitted to JOSS

---

The package is valuable, well-documented, and easy to use and I commend the maintainers for further improvements to this package since I gave them a friendly review of it a while back.
Having this functionality available in Python is a major contribution to the field.

There are a few checkboxes that still need to be addressed (plus a couple of things for the maintainers to optionally consider):

- None of the recommended badges are present in the README.md file.
- README lacks citation information
- The CONTRIBUTING.md file indicates that the package isn't taking external PRs and to open an issue if you want something. I can see why NEON might need to manage expectations/contributions, but I wonder if a more welcoming version might be something like "If you are interested in contributing changes to the codebase, please open an issue for discussion prior to submitting a pull request." This would let the maintainers moderate contributions without closing the door as firmly.
- There is currently no docs url in the pyproject.toml file. I've opened a PR that addresses this along with some other minor docs issues.
- In the portion of the 'Download and Explore NEON Data' tutorial that involves manually downloading data, I would personally find it more straightforward if the code assumed the zip file had been downloaded to the working directory, but this is likely a matter of personal preference and so just something for the maintainers to consider.
- In the 'Explore isotope data by species' section of the 'Download and Explore NEON Data' tutorial, the `plt.show()` command is missing to display the plot.
- The flake8 checks included in the CI show lots of minor cleanup most of which could be automatically cleaned up with flake8

I have submitted a PR that makes a handful of small fixes to documentation and metadata aspects of the package:
https://github.com/NEONScience/NEON-utilities-python/pull/16
