### 2025-03-26 #213 @benjamindonnachie
Estimated hours spent reviewing: Seven.

---

I concur with @ethanwhite's comments above.  Overall, I'm impressed by the level of comments in the code and found it easy to follow.

- There's lots of great input validation before sending to the API.  However, if the API ever changes, I wonder if this might be difficult to maintain.  I appreciate you may wish to minimise calls to the service, but it might be less of a maintenance overhead to return errors from the API instead.

- Console output has the potential to be confusing if included as part of a bigger program - for example, rather than "Finding available files", "Stacking data files", "Downloading files" etc., whether a message slightly more descriptive, such as "Finding available NEON files" (or similar) might be more informative for users?

- I submitted a PR for a minor typo in the example file together with a suggestion for a potentially more maintainable version of convert_byte_size: https://github.com/NEONScience/NEON-utilities-python/pull/13 and https://github.com/NEONScience/NEON-utilities-python/pull/14

- requirements.txt does not list any version numbers - could be an opportunity to specify minimum versions of major libraries - https://github.com/NEONScience/NEON-utilities-python/issues/15

- Contributing.md states not accepting external pull requests and instead create an issue.  As a result, not entirely clear on what basis contributions are accepted.

- I'd be interested to learn more about the CSV files under __resources__ - is the content static?  If not, is there an opportunity to retrieve the latest version using the API instead?

- test_aop_download.py under tests references https://github.com/NEONScience/nu-python-testing, but this returns a 404 error.

- validate_year in aop_download.py uses a regex to ensure that it is within the range 2010 - 2099.  However, I think that data is only available from January 2012 onwards?  It may be better to revise this function or, alternatively, allow the API to complete the validation.

The above are minor recommendations.  The documentation is comprehensive and it's been a joy to review your code.
