### 2025-05-08 #213 @cklunch
@benjamindonnachie @ethanwhite Thank you for your very helpful reviews, and thanks for your patience for our response. We've merged in the PRs you submitted, and made a number of updates to the package and documentation:

## Comments from @ethanwhite

> None of the recommended badges are present in the README.md file.

Badges added.

> README lacks citation information

We added a CITATION.cff file, and linked to it and to the general NEON citation guidance page from the readme.

> The CONTRIBUTING.md file indicates that the package isn't taking external PRs and to open an issue if you want something. I can see why NEON might need to manage expectations/contributions, but I wonder if a more welcoming version might be something like "If you are interested in contributing changes to the codebase, please open an issue for discussion prior to submitting a pull request." This would let the maintainers moderate contributions without closing the door as firmly.

Thanks, we've updated the CONTRIBUTING file using your suggested language.

> In the portion of the 'Download and Explore NEON Data' tutorial that involves manually downloading data, I would personally find it more straightforward if the code assumed the zip file had been downloaded to the working directory, but this is likely a matter of personal preference and so just something for the maintainers to consider.

This is a good point, NEON tutorials are currently a bit inconsistent about downloading to the working directory vs elsewhere, and I agree the working directory is simpler. I'm anticipating a broad update to tutorials in the fall, I'll see if we can standardize this at that point.

> In the 'Explore isotope data by species' section of the 'Download and Explore NEON Data' tutorial, the plt.show() command is missing to display the plot.

Fixed!

> The flake8 checks included in the CI show lots of minor cleanup most of which could be automatically cleaned up with flake8

Done, minor issues resolved.

## Comments from @benjamindonnachie

> There's lots of great input validation before sending to the API. However, if the API ever changes, I wonder if this might be difficult to maintain. I appreciate you may wish to minimise calls to the service, but it might be less of a maintenance overhead to return errors from the API instead.

Interesting point. The NEON API has been very stable, so this hasn't been an issue with the R version of this package, but definitely something we'll keep in mind as we make updates.

> Console output has the potential to be confusing if included as part of a bigger program - for example, rather than "Finding available files", "Stacking data files", "Downloading files" etc., whether a message slightly more descriptive, such as "Finding available NEON files" (or similar) might be more informative for users?

We've added to these messages, specifying "NEON" in many places and the specific data product ID where relevant.

> requirements.txt does not list any version numbers - could be an opportunity to specify minimum versions of major libraries - https://github.com/NEONScience/NEON-utilities-python/issues/15

Added version numbers to requirements.txt

> Contributing.md states not accepting external pull requests and instead create an issue. As a result, not entirely clear on what basis contributions are accepted.

Text is updated using the language Ethan recommended.

> I'd be interested to learn more about the CSV files under resources - is the content static? If not, is there an opportunity to retrieve the latest version using the API instead?

We removed one file that wasn't used. The others are unfortunately not available via the API - we would prefer to get them there if we could! But, they are very rarely updated.

> test_aop_download.py under tests references https://github.com/NEONScience/nu-python-testing, but this returns a 404 error.

Fixed, removed pointer to internal repo.

> validate_year in aop_download.py uses a regex to ensure that it is within the range 2010 - 2099. However, I think that data is only available from January 2012 onwards? It may be better to revise this function or, alternatively, allow the API to complete the validation.

Regex updated.
