### 2025-05-21 #213 @benjamindonnachie
I concur, amendments look good!  Since the API has remained stable, validation in the application shouldn't be an issue for now, but certainly if there are changes, it is worth considering moving validation to the api - although I accept there may be a desire to trade off maintainability with reduced traffic.

Approved - I will update my checklist now.  Thanks again for submitting this package for review!
