### 2025-05-23 #213 @JuliMillan
---

🎉 Neonutilities has been approved by pyOpenSci! Thank you @cklunch for submitting neonutilities and many thanks to @ethanwhite and @benjamindonnachie for reviewing this package! 😸

- [x] Add the badge for pyOpenSci peer-review to the README.md of neonutilities. The badge should be `[![pyOpenSci Peer-Reviewed](https://pyopensci.org/badges/peer-reviewed.svg)](https://github.com/pyOpenSci/software-review/issues/issue-number)`.

---
