### 2025-05-27 #213 @cklunch
@ethanwhite @benjamindonnachie Thank you so much!

@JuliMillan I can't check the boxes in your comment, but I've activated the Zenodo watching and DOI, and added the DOI and pyOpenSci badges to the repo. I filled out the survey. Let me know if there are any other actions needed!
